log more information
