+++ GestureSign.ControlPanel/GestureSign.ControlPanel.csproj
@@ -97,6 +97,7 @@
     </Reference>
     <Reference Include="System" />
     <Reference Include="System.Drawing" />
+    <Reference Include="System.Management" />
     <Reference Include="System.Windows.Forms" />
     <Reference Include="System.Windows.Interactivity, Version=4.5.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35, processorArchitecture=MSIL">
       <HintPath>..\packages\MahApps.Metro.1.2.4.0\lib\net45\System.Windows.Interactivity.dll</HintPath>
@@ -118,7 +119,7 @@
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </ApplicationDefinition>
-    <Compile Include="Common\Net.cs" />
+    <Compile Include="Common\ErrorReport.cs" />
     <Compile Include="Converters\AllowSingleBoolConverter.cs" />
     <Compile Include="Converters\SelectedIndex2BoolConverter.cs" />
     <Compile Include="Converters\SelectedIndex2VisibilityConverter.cs" />
